+++ SandSim/StoneCell.cpp
@@ -1,6 +1,6 @@
 #include "StoneCell.h"
 
-StoneCell::StoneCell(size_t x, size_t y) : Cell(x, y, CellType::Stone) {}
+StoneCell::StoneCell(int x, int y) : Cell(x, y, CellType::Stone) {}
 
 void StoneCell::update(std::vector<std::vector<std::shared_ptr<Cell>>>& grid) {
