+++ SandSim/WaterCell.h
@@ -3,6 +3,6 @@
 
 class WaterCell : public Cell {
 public:
-    WaterCell(size_t x, size_t y);
+    WaterCell(int x, int y);
     void update(std::vector<std::vector<std::shared_ptr<Cell>>>& grid) override;
 };
