+++ SandSim/WoodCell.h
@@ -3,6 +3,6 @@
 
 class WoodCell : public Cell {
 public:
-    WoodCell(size_t x, size_t y);
+    WoodCell(int x, int y);
     void update(std::vector<std::vector<std::shared_ptr<Cell>>>& grid) override;
 };
